Scope source files to component pack
