--- lib/weird_phlex/component_pack/variant.rb
@@ -1,56 +1,0 @@
-# frozen_string_literal: true
-
-require 'open3'
-
-module WeirdPhlex
-  class ComponentPack
-    class Variant
-      LIBRARAY_AND_VARIANT_REGEX = /\Aweird_phlex_pack-(?<library>[_\w]+)-(?<variant>[_\w]+)\Z/
-
-      def self.all
-        # We assume that both `weird_phlex` and all component packs are loaded in Bundler group `:development`
-        # and are thus available at the same time.
-        Gem.loaded_specs
-          .select { |name, _gem_specification| name.match(LIBRARAY_AND_VARIANT_REGEX) }
-          .values
-          .map { new(_1) }
-      end
-
-      def initialize(gem_specification)
-        @name = gem_specification.name
-        @gem = gem_specification.name
-        @gem_path = Pathname.new(gem_specification.gem_dir)
-        @library, @variant = library_and_variant(gem_specification.name)
-        @component_path = @gem_path.join('lib', 'weird_phlex_pack', @library, @variant)
-      end
-
-      def components
-        files
-          .select(&:component_file?)
-          .group_by(&:component)
-          .map do |component_name, files|
-            WeirdPhlex::ComponentPack::Component.new(component_name, files: files)
-          end
-      end
-
-      def files
-        file_paths.map { WeirdPhlex::ComponentPack::File.new(_1, component_path: @component_path) }
-      end
-
-      private
-
-      # Potentially use Gem::Specification.lib_files
-      def file_paths
-        Dir['**/*', base: @component_path.to_s].map { @component_path.join(_1) }.select(&:file?)
-      end
-
-      def library_and_variant(name)
-        if (matches = name.match(LIBRARAY_AND_VARIANT_REGEX))
-          [matches[:library], matches[:variant]]
-        else
-          raise "Regex error: Could not parse component pack name '#{name}'!"
-        end
-      end
-    end
-  end
-end
